requirements.txt
Flask==3.1.3
opencv-python==4.13.0.92
mediapipe==0.10.9
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
pyspellchecker==0.8.1